# src/main/java/com/bigobooks/dto/StockQuantityRequestDto.java
package com.bigobooks.dto;

import jakarta.validation.constraints.NotNull;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class StockQuantityRequestDto {

    @NotNull
    private Integer quantity;
}